Make BaxH5Reader more robust to region table screwups
This was another place where the primary pipeline was outputing a
regiontable that I would consider "broken".  Previously (bug 23585) we
saw a bug where a region table for a zmw could be missing an hqregion;
here we see another problem, where a sequencing ZMW is lacking a
region table entry altogether.

This fix makes pbcore more robust, but primary should not be
outputting busted region tables like this.  I will follow up with John
N.

[Bug 25273: "Meredith's BasH5IO issue"]

[git-p4: depot-paths = "//depot/software/bioinformatics/lib/python/pbcore/": change = 137152]